Move non-negative number helpers to abstract-ops
