+++ src/lib/abstract-ops/miscellaneous.ts
@@ -1,0 +1,29 @@
+import NumberIsNaN from '../../stub/number-isnan';
+
+export function IsFiniteNonNegativeNumber(v: number): boolean {
+  if (IsNonNegativeNumber(v) === false) {
+    return false;
+  }
+
+  if (v === Infinity) {
+    return false;
+  }
+
+  return true;
+}
+
+export function IsNonNegativeNumber(v: number): boolean {
+  if (typeof v !== 'number') {
+    return false;
+  }
+
+  if (NumberIsNaN(v)) {
+    return false;
+  }
+
+  if (v < 0) {
+    return false;
+  }
+
+  return true;
+}
